changed min and max to uint16_t

+++ libraries/Servo/src/Servo.h
@@ -111,8 +111,8 @@ public:
   bool attached();                   // return true if this servo is attached, otherwise false 
 private:
    uint8_t servoIndex;               // index into the channel data for this servo
-   int8_t min;                       // minimum is this value times 4 added to MIN_PULSE_WIDTH    
-   int8_t max;                       // maximum is this value times 4 added to MAX_PULSE_WIDTH   
+   int16_t min;                       // minimum pulse in µs    
+   int16_t max;                       // maximum pulse in µs   
 };
 
 #endif
